login dan crud

+++ masyarakat/show_pengaduan.php
@@ -1,5 +1,5 @@
 <?php 
-	require 'koneksi.php';
+	require '../koneksi.php';
 	$pengaduan = mysqli_query($koneksi, "SELECT * FROM pengaduan INNER JOIN masyarakat ON pengaduan.nik = masyarakat.nik");
 ?>
 
@@ -11,6 +11,8 @@
 	<title>Daftar Pengaduan</title>
 </head>
 <body>
+	<?php include 'sidebar.php'; ?>
+
 	<a href="insert_pengaduan.php">Tambah Pengaduan</a>
 	<table border="1" cellpadding="10" cellspacing="0">
 		<thead>
@@ -32,11 +34,15 @@
 					<td><?= $dataPengaduan['tgl_pengaduan']; ?></td>
 					<td><?= $dataPengaduan['nik']; ?></td>
 					<td><?= $dataPengaduan['isi_laporan']; ?></td>
-					<td><img style="width: 150px" src="img/<?= $dataPengaduan['foto']; ?>"></td>
+					<td><img style="width: 150px" src="../img/<?= $dataPengaduan['foto']; ?>"></td>
 					<td><?= $dataPengaduan['status']; ?></td>
 					<td>
-						<a href="update_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Ubah</a>
-						<a onclick="return confirm('Apakah anda yakin menghapus data pengaduan dengan isi laporan <?= $dataPengaduan['isi_laporan']; ?>?')" href="delete_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Hapus</a>
+						<?php if ($dataPengaduan['status'] == 'proses'): ?>
+							<a href="update_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Ubah</a>
+							<a onclick="return confirm('Apakah anda yakin menghapus data pengaduan dengan isi laporan <?= $dataPengaduan['isi_laporan']; ?>?')" href="delete_pengaduan.php?id_pengaduan=<?= $dataPengaduan['id_pengaduan']; ?>">Hapus</a>
+						<?php else: ?>
+							<span>selesai</span>
+						<?php endif ?>
 					</td>
 				</tr>
 			<?php endforeach ?>
